+++ sdk/appcenter-distribute/src/main/java/com/microsoft/appcenter/distribute/ReleaseDownloadListener.java
@@ -10,6 +10,7 @@ import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
 import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
 import android.support.annotation.UiThread;
 import android.support.annotation.WorkerThread;
 import android.widget.Toast;
@@ -112,8 +113,9 @@ class ReleaseDownloadListener implements ReleaseDownloader.Listener {
 
     @WorkerThread
     @Override
-    public void onError(@NonNull String errorMessage) {
-        AppCenterLog.error(LOG_TAG, errorMessage);
+    public void onError(@Nullable String errorMessage) {
+        AppCenterLog.error(LOG_TAG, String.format(Locale.ENGLISH, "Failed to download %s (%d) update: %s",
+                mReleaseDetails.getShortVersion(), mReleaseDetails.getVersion(), errorMessage));
         HandlerUtils.runOnUiThread(new Runnable() {
 
             @Override