+++ sdk/appcenter-distribute/src/main/java/com/microsoft/appcenter/distribute/download/ReleaseDownloader.java
@@ -8,6 +8,7 @@ package com.microsoft.appcenter.distribute.download;
 import android.net.Uri;
 import android.support.annotation.AnyThread;
 import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
 import android.support.annotation.WorkerThread;
 
 import com.microsoft.appcenter.distribute.ReleaseDetails;
@@ -83,6 +84,6 @@ public interface ReleaseDownloader {
          *
          * @param errorMessage The message of the exception.
          */
-        void onError(String errorMessage);
+        void onError(@Nullable String errorMessage);
     }
 }